Change execute arg to list and refactor demo classes to own package
