+++ src/main/java/demo/CreateIndexWorkflowStep.java
@@ -6,7 +6,7 @@
  * this file be licensed under the Apache-2.0 license or a
  * compatible open source license.
  */
-package org.opensearch.flowframework.template;
+package demo;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -15,6 +15,7 @@ import org.opensearch.flowframework.workflow.WorkflowData;
 import org.opensearch.flowframework.workflow.WorkflowInputData;
 import org.opensearch.flowframework.workflow.WorkflowStep;
 
+import java.util.List;
 import java.util.Map;
 import java.util.concurrent.CompletableFuture;
 
@@ -30,18 +31,24 @@ public class CreateIndexWorkflowStep implements WorkflowStep {
     }
 
     @Override
-    public CompletableFuture<WorkflowData> execute(WorkflowData... data) {
-        logger.debug("Executing with " + data.length + " args");
+    public CompletableFuture<WorkflowData> execute(List<WorkflowData> data) {
+        logger.debug("Executing with {} input args.", data.size());
         CompletableFuture<WorkflowData> future = new CompletableFuture<>();
         CompletableFuture.runAsync(() -> {
             if (inputData != null) {
-                Map<String, String> params = inputData.getParams();
-                Map<String, Object> content = inputData.getContent();
-                logger.debug("Initialized params: {}, content: {}", params, content);
+                logger.debug("Initialized params: {}, content: {}", inputData.getParams(), inputData.getContent());
             }
             for (WorkflowData wfData : data) {
                 Map<String, Object> content = wfData.getContent();
-                logger.debug("Received content from previous step: {}", content);
+                if (wfData instanceof WorkflowInputData) {
+                    logger.debug(
+                        "Previous step sent params: {}, content: {}",
+                        ((WorkflowInputData) wfData).getParams(),
+                        wfData.getContent()
+                    );
+                } else {
+                    logger.debug("Received from previous step: content: {}", content);
+                }
             }
             // do some work
             try {
