+++ src/main/java/demo/DataDemo.java
@@ -6,10 +6,12 @@
  * this file be licensed under the Apache-2.0 license or a
  * compatible open source license.
  */
-package org.opensearch.flowframework.template;
+package demo;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import org.opensearch.flowframework.template.ProcessNode;
+import org.opensearch.flowframework.template.TemplateParser;
 import org.opensearch.flowframework.workflow.WorkflowStep;
 
 import java.io.IOException;
@@ -26,18 +28,16 @@ import java.util.stream.Collectors;
 /**
  * Demo class exercising {@link TemplateParser}. This will be moved to a unit test.
  */
-public class Demo {
+public class DataDemo {
 
-    private static final Logger logger = LogManager.getLogger(Demo.class);
+    private static final Logger logger = LogManager.getLogger(DataDemo.class);
 
     // This is temporary. We need a factory class to generate these workflow steps
     // based on a field in the JSON.
     private static Map<String, WorkflowStep> workflowMap = new HashMap<>();
     static {
-        workflowMap.put("fetch_model", new DemoWorkflowStep(3000));
-        workflowMap.put("create_ingest_pipeline", new DemoWorkflowStep(3000));
-        workflowMap.put("create_search_pipeline", new DemoWorkflowStep(5000));
-        workflowMap.put("create_neural_search_index", new DemoWorkflowStep(2000));
+        workflowMap.put("create_index", new CreateIndexWorkflowStep());
+        workflowMap.put("create_another_index", new CreateIndexWorkflowStep());
     }
 
     /**
@@ -46,7 +46,7 @@ public class Demo {
      * @param args unused
      */
     public static void main(String[] args) {
-        String path = "src/test/resources/template/demo.json";
+        String path = "src/test/resources/template/datademo.json";
         String json;
         try {
             json = new String(Files.readAllBytes(Paths.get(path)));
@@ -71,7 +71,6 @@ public class Demo {
                         predecessors.stream().map(p -> p.id()).collect(Collectors.joining(", "))
                     )
             );
-            // TODO need to handle this better, passing an argument when we start them all at the beginning is silly
             futureList.add(n.execute());
         }
         futureList.forEach(CompletableFuture::join);
